refactor: 자동차 이름 검증 로직 클래스 분리
